Added InvokeScript and InvokePowershell methods to DocumentRepository
This commit includes the addition of two new methods to the DocumentRepository class: InvokeScript and InvokePowershell. InvokeScript uses reflection to dynamically invoke methods based on attributes, specifically looking for methods with the ScriptTypeAttribute and a type of "Powershell". InvokePowershell is a private static method that executes a script using a new process. Both methods have been decorated with SuppressMessage attributes to ignore specific code analysis warnings. The using directives have been updated to include additional namespaces and a comment block has been added at the top of the file to track changes. The NotImplementedException in the InvokeScript method has been removed as the method is now fully implemented.