Updated methods and simplified controller addition
In `AutomationExtensions.cs`, a new method `GetClickablePoint` was added to get a clickable point on the UI Automation element based on the specified alignment and offsets. In `User32Controller.cs`, the method of retrieving sessions was changed to use `_domain.SessionsRepository.GetSession(session).Session` instead of `_domain.GetSession(session)`. The `GetClickablePoint` and `SendNativeClick` methods were updated to no longer take `sessionModel.ScaleRatio` as a parameter. Lastly, in `Program.cs`, the addition of the `TextPlainInputFormatter` to controllers was simplified and is now directly added in the `AddControllers` method call.