Replace dotnet-desktop.yml with GithubActions.yml and update Uia.sln
Replaced `dotnet-desktop.yml` with a new GitHub Actions workflow file `GithubActions.yml` for building and releasing the `Uia.DriverServer` project. The workflow is triggered on push to the main branch and includes jobs for version calculation, build and release creation. Updated `Uia.sln` to change the project type of `Uia.DriverServer.Formatters` and added new solution items `.git` and `.documentation`.