Update logging level in appsettings.json
Updated the logging level for "Default" and "Microsoft.AspNetCore" in the "Logging" section of the appsettings.json file from "Error" to "Information". This change will allow the system to log informational messages in addition to error messages, aiding in debugging and understanding the application flow. However, it may also lead to an increase in log data generation.