Refactor UiaOptions and update session creation methods
This commit includes a major refactor of the `UiaOptions` class, which has been replaced with `UiaOptionsModel`. This new class includes an additional `App` property representing the application path. The `NewSession` method in `SessionsRepository.cs` has been updated to handle these changes, with adjustments to the assignment of `matchCapabilities` and `options` variables. The checks for `isApp` and `isDesktop` now rely on the `App` property of `options`. The `NewApplicationSession` method now takes `UiaOptionsModel` and `IDictionary<string, object>` as parameters, with the `fileName` variable assigned the value of `App` from `options`. Lastly, the logging of the session creation process has been removed.