Update GitHub Actions for preview releases
Changed the `artifactType` in the GitHub Actions workflow from `Production` to `Preview`, indicating a shift towards generating artifacts for testing or preview rather than for final production use. Additionally, updated the GitHub release configuration from draft to prerelease, reflecting the intention for these artifacts to be used for pre-final testing or preview purposes.