Add Emgu configuration to pipeline YAML files
Implemented a matrix strategy in `preview-pipeline.yml` and `release-pipeline.yml` to include a new `Release_Emgu` configuration with `useEmgu` set to `true`. Added a step to update the `<UseEmgu>` node in the project file based on the matrix variable, along with logging for confirmation. Introduced a new step for setting Emgu.CV dependencies, initializing version variables for the pipeline.